renamed long names: lint fail fixed
